[LC-335] - Addition of new @ModelAttribute TokenRequest caused the BindingResult to contain the binding result for the TokenRequest rather than the exected SignupForm.
Spring will put the binding result of a model attribute in to the first available BindingResult in the method signature after the model attribute, in this case the method signature was "(@ModelAttribute signupFrom, @ModelAttribute tokenRequest, BindingResult)", this would cause Spring to bind signupFrom and write the results to the BindingResult but then bind tokenRequest and reuse the BindingResult overwirtting the result for SignupFrom.

Added a new BindingResult immediately after the model attribute for signupForm, this will cause Spring to write the binding results of the attributes to two different BindingResults.

Removed the standalone SignupFormValidator in favour of using JSR 380 validation annotations on the SignupForm model class.

Fixed associated unit tests for above changes.