LC-337: change email updates for agency self signup
Update change email verify endpoints to handle invited, whitelisted and agency users.

The email/verify method now handles the following scenarios of email change:
- whitelisted/invited to whitelisted
- whitelisted/invited to agency
- agency to agency
- agency to whitelisted

If a new email is whitelisted or invited, the user should continue to verify the change of email upon clicking the link in the verify email as normal.

If the new email is not whitelisted and is agency, they will now be required to verify their token credentials with domain, organisation and token values.
This token verification was previously done post-login, but by putting as part of the overall change email verifification flow, we have greater control of how the user move through the change email journey, i.e. they shouldnt be able to change their email to an agency token domain if they cannot submit verified token credentials.

Improving the email updated success page to present the user's new email as well as a link to logout and in again.

Updates to controllers and services, removing some redundant classes and methods now that we have simplfied the agency signup process.