LC-337: remove email recently updated flag
With the updates to signup and change email, the emailRecentlyUpdated flag is now redundant. Removing this field from domains, services, controllers and migrations.

Removing the SpringUserUtils class because the emailRecentlyUpdated flag is no longer used to determine whether the user should be presented with the enterToken screen to validate an agency token.