LC-332: update reactivation process
With updates for agency self-signup, the account reactivation journey needs to be updated so that the user can be forced through the 'verify token' flow.
Adding new migrations, domains, services and controllers to allow a user to complete the reactivation upon clicking a link in the reactivation email (which will be initiated by an admin user on IDM.

Made the AgencyTokenVerification controllers, forms and templates a lot more generic, so the controllers and services can be used for both changing an email and reactivating an account.

Adding a VerificationCodeTypeDetermination service class to determine whether the user is coming from an email update verification email or an account reactivation verification email, thus allowing the agnecy token verification controller to be generic and call the appropriate service class accordingly.

- replaced string literals for templates and redirects with constants
- remove/add agencyTokenUid field for Identity when deactivating and reactivating
- database migrations for new Reactivation object
- updates to templates
- unit tests